requirements.txt
google-adk
google-genai
mcp[cli]
requests
fastapi
uvicorn